drivers/modem/alt1250: Revise the power control sequence
Fixed based on the following two specifications.

1. Do not assert if do not use PMU_POWER_BUTTON.

2. The PMU_SHUTDOWN pin is used to initiate a power cycle
that resets the device. It should be asserted for a duration
of >100usec to perform a hard reset.

The past reset sequence used the power off and power on
interfaces. This operation is not a correct reset sequence,
so the reset interface was added.