drivers/modem/alt1250: Fix a bug that name resolution fails
When the application stopped netinfo events with
lte_set_report_netinfo(), it sent a command to ALT1250
to stop netinfo events. As a result, the DNS address
information contained in netinfo could not be obtained,
resulting in address resolution failure.
The command to stop netinfo event should not be sent.